mv should only prompt for file being unwriteable when stdin is a tty.

+++ toys/posix/cp.c
@@ -432,7 +432,7 @@ void cp_main(void)
         // Prompt if -i or file isn't writable.  Technically "is writable" is
         // more complicated (022 is not writeable by the owner, just everybody
         // _else_) but I don't care.
-        if (exists && (FLAG(i) || !(st.st_mode & 0222))) {
+        if (exists && (FLAG(i) || (!(st.st_mode & 0222) && isatty(0)))) {
           fprintf(stderr, "%s: overwrite '%s'", toys.which->name, TT.destname);
           if (!yesno(0)) rc = 0;
           else unlink(TT.destname);
